Refactored folder structure for routes
