+++ src/routes/AddPage.jsx
@@ -1,11 +1,11 @@
 import React, { useState, useContext } from "react";
 import { v4 as uuidv4 } from "uuid";
 
-import Button from "../../views/Button";
-import Chip from "../../views/Chip";
-import { AddIcon, CloseIcon } from "../../views/Icons";
+import Button from "../views/Button";
+import Chip from "../views/Chip";
+import { AddIcon, CloseIcon } from "../views/Icons";
 
-import { FirebaseContext } from "../../Firebase";
+import { FirebaseContext } from "../Firebase";
 import { motion } from "framer-motion";
 
 const AddPage = ({ closeFunction }) => {
