+++ src/routes/ItemCard.jsx
@@ -1,12 +1,12 @@
-import Chip from "../../views/Chip";
-import Button from "../../views/Button";
-import { FavouritesIcon, ShoppingIcon } from "../../views/Icons";
+import Chip from "../views/Chip";
+import Button from "../views/Button";
+import { FavouritesIcon, ShoppingIcon } from "../views/Icons";
 
 import _ from "lodash";
 import { motion } from "framer-motion";
 
-import noimg from "../../assets/no-img.png";
-import { AppContext } from "../../Context";
+import noimg from "../assets/no-img.png";
+import { AppContext } from "../Context";
 import { useContext } from "react";
 
 const ItemCard = ({ item, index }) => {
