+++ src/routes/SignIn.jsx
@@ -1,8 +1,8 @@
-import Logo from "../../views/Logo";
-import Button from "../../views/Button";
+import Logo from "../views/Logo";
+import Button from "../views/Button";
 import { useContext } from "react";
-import { FirebaseContext } from "../../Firebase";
-import { GoogleIcon } from "../../views/Icons";
+import { FirebaseContext } from "../Firebase";
+import { GoogleIcon } from "../views/Icons";
 
 const SignInPage = () => {
   const { signInGoogle } = useContext(FirebaseContext);
